SA-1850 - Extract delay reason table from DelayReasonsTab
