+++ src/pages/reportBuilder/components/tables/index.js
@@ -1,1 +1,2 @@
 export { default as BounceReasonTable } from './BounceReasonTable';
+export { default as DelayReasonTable } from './DelayReasonTable';
