+++ src/pages/reportBuilder/components/tabs/DelayReasonsTab.js
@@ -1,36 +1,11 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
-import { LongTextContainer, Percent, TableCollection } from 'src/components';
 import { refreshDelayReportV2 as refreshDelayReport } from 'src/actions/delayReport';
-import { safeRate } from 'src/helpers/math';
-import {
-  EmptyWrapper,
-  FilterBoxWrapper,
-  LoadingWrapper,
-  TableCollectionBody,
-  TableWrapper,
-} from '../Wrappers';
 import { useReportBuilderContext } from '../../context/ReportBuilderContext';
+import { DelayReasonTable } from '../tables';
 
-const filterBoxConfig = {
-  show: true,
-  itemToStringKeys: ['domain', 'reason'],
-  exampleModifiers: ['domain', 'reason'],
-  matchThreshold: 5,
-  label: 'Filter',
-  wrapper: FilterBoxWrapper,
-};
-
-const columns = [
-  { label: 'Total Delays', sortKey: 'count_delayed' },
-  { label: 'Delayed First Attempt (%)', sortKey: 'count_delayed_first' },
-  { label: 'Reason', sortKey: 'reason', width: '45%' },
-  { label: 'Domain', sortKey: 'domain' },
-];
-
-export function DelayReasonsTab(props) {
+export function DelayReasonsTab({ reasons, totalAccepted, refreshDelayReport, tableLoading }) {
   const { state: reportOptions } = useReportBuilderContext();
-  const { loading, reasons, totalAccepted, refreshDelayReport } = props;
 
   useEffect(() => {
     if (reportOptions.to && reportOptions.from) {
@@ -38,45 +13,15 @@ export function DelayReasonsTab(props) {
     }
   }, [refreshDelayReport, reportOptions]);
 
-  const getRowData = rowData => {
-    const { reason, domain, count_delayed, count_delayed_first } = rowData;
-    return [
-      count_delayed,
-      <span>
-        {count_delayed_first} (<Percent value={safeRate(count_delayed_first, totalAccepted)} />)
-      </span>,
-      <LongTextContainer text={reason} />,
-      domain,
-    ];
-  };
-
-  if (loading) {
-    return <LoadingWrapper />;
-  }
-
-  if (!reasons.length) {
-    return <EmptyWrapper message="No delay reasons to report" />;
-  }
   return (
-    <TableCollection
-      columns={columns}
-      rows={reasons}
-      getRowData={getRowData}
-      pagination
-      defaultSortColumn="count_delayed"
-      defaultSortDirection="desc"
-      wrapperComponent={TableWrapper}
-      filterBox={filterBoxConfig}
-    >
-      {props => <TableCollectionBody {...props} />}
-    </TableCollection>
+    <DelayReasonTable reasons={reasons} totalAccepted={totalAccepted} loading={tableLoading} />
   );
 }
 
 const mapStateToProps = state => {
   const { aggregates } = state.delayReport;
   return {
-    loading: state.delayReport.aggregatesLoading || state.delayReport.reasonsLoading,
+    tableLoading: state.delayReport.aggregatesLoading || state.delayReport.reasonsLoading,
     reasons: state.delayReport.reasons,
     totalAccepted: aggregates ? aggregates.count_accepted : 1,
   };
